+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Added `Print::global()` and `bullet_stream::global::set_writer`. Use these to preserve the newline indentation when handling dropped structs or errors ()
+
 ## v0.3.0 - 2024/08/14
 
 - Added `bullet_stream::strip_ansi` (https://github.com/schneems/bullet_stream/pull/11)