+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Added `bullet_stream::global::print` functions for writing formatted output without needing to preserve state ()
 - Added `Print::global()` and `bullet_stream::global::set_writer`. Use these to preserve the newline indentation when handling dropped structs or errors ()
 
 ## v0.3.0 - 2024/08/14